fix: remove extra code from AppDelegate
