+++ Project 15 -  AnimatedSplash/AnimatedSplash/AppDelegate.swift
@@ -9,58 +9,23 @@
 import UIKit
 
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate, CAAnimationDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
-    private var mask: CALayer?
-    private var imageView: UIImageView!
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
 
         window = UIWindow(frame: UIScreen.main.bounds)
         window?.rootViewController = ViewController()
 
-        imageView = UIImageView(frame: window!.frame)
-        imageView.image = UIImage(named: "twitterscreen")
-        window!.addSubview(imageView)
+        let maskImage = LaunchView(frame: window!.frame)
+        window?.addSubview(maskImage)
 
-        mask = CALayer()
-        if let mask = mask {
-            mask.contents = UIImage(named: "twitter")?.cgImage
-            mask.contentsGravity = kCAGravityResizeAspect
-            mask.bounds = CGRect(x: 0, y: 0, width: 100, height: 81)
-            mask.position = CGPoint(x: imageView.frame.size.width / 2, y: imageView.frame.size.height / 2)
-            imageView.layer.mask = mask
-            animateMask()
-        }
         window?.backgroundColor = UIColor(red: 0.117, green: 0.631, blue: 0.949, alpha: 1)
         window?.makeKeyAndVisible()
         UIApplication.shared.isStatusBarHidden = true
         
         return true
-    }
-
-    private func animateMask() {
-        let keyFrameAnimation = CAKeyframeAnimation(keyPath: "bounds")
-        keyFrameAnimation.delegate = self
-        keyFrameAnimation.duration = 0.6
-        keyFrameAnimation.beginTime = CACurrentMediaTime() + 0.5
-        keyFrameAnimation.timingFunctions = [
-            CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseOut),
-            CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseOut)
-        ]
-        let initalBounds = NSValue(cgRect: mask!.bounds)
-        let secondBounds = NSValue(cgRect: CGRect(x: 0, y: 0, width: 90, height: 73))
-        let finalBounds = NSValue(cgRect: CGRect(x: 0, y: 0, width: 1600, height: 1300))
-        keyFrameAnimation.values = [initalBounds, secondBounds, finalBounds]
-        keyFrameAnimation.keyTimes = [0, 0.3, 1]
-        mask!.add(keyFrameAnimation, forKey: "bounds")
-    }
-
-    func animationDidStop(_ anim: CAAnimation, finished flag: Bool) {
-        if flag == true {
-            imageView.layer.mask = nil
-        }
     }
 
     func applicationWillResignActive(_ application: UIApplication) {
